refactor(drizzle): split checkout session persistence
- move Drizzle checkout-session row loading, current-session selection, and item persistence into a repository-local helper
- keep DrizzleCheckoutSessionRepository focused on port wiring and persistence error translation
- improve the Fallow checkout-session repository hotspot without changing adapter semantics
